Move everything related to electron into its own directory
