+++ lib/api/window.js
@@ -15,7 +15,7 @@
 
 const electron = require('electron')
 const platform = require('../platform')
-const electronWindowManager = require('../electronWindowManager')
+const windowManagement = require('../electron/windowManagement.js')
 
 /**
  * A factory function
@@ -44,7 +44,7 @@ function factory (api, workspace) {
    * @returns { Promise.<Boolean> }
    */
   async function toggleMaximize () {
-    const window = await electronWindowManager.getWindowFromWorkspace(workspace.id)
+    const window = await windowManagement.getWindowFromWorkspace(workspace.id)
     if (!window) {
       return false
     }
