+++ lib/electron/electron.js
@@ -11,13 +11,13 @@ const {
   powerSaveBlocker
 } = require('electron')
 
-const paths = require('./paths')
-const platform = require('./platform')
+const paths = require('../paths')
+const platform = require('../platform')
 
-const ProjectFile = require('./ProjectFile')
-const UserDefaults = require('./UserDefaults')
-const WorkspaceRegistry = require('./WorkspaceRegistry')
-const electronWindowManager = require('./electronWindowManager')
+const ProjectFile = require('../ProjectFile')
+const UserDefaults = require('../UserDefaults')
+const windowManagement = require('./windowManagement')
+const WorkspaceRegistry = require('../WorkspaceRegistry')
 
 /**
  * The template for
@@ -132,7 +132,7 @@ as workspace if it was opened with one
    * @param { String } filePath
    */
   async function openWithFile (filePath) {
-    const alreadyOpenedWindow = await electronWindowManager.getWindowFromProjectFile(filePath)
+    const alreadyOpenedWindow = await windowManagement.getWindowFromProjectFile(filePath)
     if (alreadyOpenedWindow) {
       alreadyOpenedWindow.focus()
       return
@@ -227,7 +227,7 @@ function initWindow (url) {
     */
     e.preventDefault()
 
-    const workspace = await electronWindowManager.getWorkspaceFromWindow(window)
+    const workspace = await windowManagement.getWorkspaceFromWindow(window)
     if (workspace) {
       WorkspaceRegistry.getInstance().delete(workspace.id)
       workspace.teardown()
@@ -295,7 +295,7 @@ async function open (filePath) {
   focus that rather than opening
   a new one
   */
-  const alreadyOpenedWindow = await electronWindowManager.getWindowFromProjectFile(filePath)
+  const alreadyOpenedWindow = await windowManagement.getWindowFromProjectFile(filePath)
   if (alreadyOpenedWindow) {
     alreadyOpenedWindow.focus()
     return
@@ -317,7 +317,7 @@ async function open (filePath) {
  * @param { BrowserWindow } window
  */
 async function save (window) {
-  const workspace = await electronWindowManager.getWorkspaceFromWindow(window)
+  const workspace = await windowManagement.getWorkspaceFromWindow(window)
   if (!workspace) {
     return
   }
@@ -342,7 +342,7 @@ async function save (window) {
  * @param { BrowserWindow } window
  */
 async function saveAs (window) {
-  const workspace = await electronWindowManager.getWorkspaceFromWindow(window)
+  const workspace = await windowManagement.getWorkspaceFromWindow(window)
   if (!workspace) {
     return
   }
