+++ lib/electron/windowManagement.js
@@ -3,9 +3,9 @@
 // SPDX-License-Identifier: MIT
 
 const { BrowserWindow } = require('electron')
-const utils = require('./utils')
+const utils = require('../utils')
 
-const WorkspaceRegistry = require('./WorkspaceRegistry')
+const WorkspaceRegistry = require('../WorkspaceRegistry')
 
 /**
  * Get the window currently hosting
