+++ lib/init-electron.js
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: MIT
 
 const fs = require('node:fs')
-const electron = require('./electron')
+const electron = require('./electron/electron')
 
 const paths = require('./paths')
 const UserDefaults = require('./UserDefaults')
